WebTerminal app structure is now consistent with OIDE standards.
