--- oide/apps/webterminal/handlers.py
@@ -1,10 +1,0 @@
-import tornado.web
-
-
-import oide.lib.decorators
-import oide.apps.webterminal.settings as app_settings
-from oide.lib.handlers.base import BaseHandler
-
-
-
-
