+++ oide/apps/webterminal/settings.py
@@ -6,19 +6,16 @@ APP_SPECIFICATION = {
     'APP_DESCRIPTION': {
         'name': 'Terminal',
         'link': '/#/terminal',
-        'description': 'The ShellInABox web terminal.'
+        'description': 'An easy-to-use web-based SSH terminal.'
     },
     'NG_MODULE_NAME': 'terminal',
     'NG_MODULE_STYLESHEETS': (
         'terminal.css',
     ),
     'NG_MODULE_SCRIPTS': (
-        'terminado.js',
         'terminal.js',
     ),
 }
-
-SHELLINABOX_URL = 'http://localhost:10443'
 
 try:
     from local_settings import *
