+++ oide/apps/webterminal/static/terminal.css
@@ -20,3 +20,8 @@
   background-color: white !important;
   border: none !important;
 }
+
+.terminal-cursor {
+  color: #93c54b;
+  background-color: #93c54b;
+}
